requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
pandas==2.0.3
openpyxl==3.1.2
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
email-validator==2.0.0
schedule==1.2.0
Werkzeug==2.3.7